requirements.txt
streamlit==1.31.0
python-docx==1.0.1
openai==1.66.3
pandas==2.1.4
PyYAML==6.0.1
Pillow==10.2.0
python-dotenv==1.0.0 